refactoring controllers and routes
